Add proper groupings to timeline
